consolidate socket files in a /socket folder
