+++ socket/socketHandler.js
@@ -1,4 +1,4 @@
-import Actions from "../../types/Actions";
+import Actions from "../types/Actions";
 import {
   handleGameStart,
   handleMoveMade,
