+++ views/gameView/onlineGameView/OnlineGameView.jsx
@@ -15,8 +15,8 @@ import Board from "../../../components/Board/Board.jsx";
 import { useState, useEffect } from "react";
 import getStartingBoard from "../../../components/Board/board.js";
 import PlayerCard from "../../../components/Board/PlayerCard.jsx";
-import createSocket from "../../../components/websocket.js";
-import socketHandler from "../../../components/Board/socketHandler.js";
+import createSocket from "../../../socket/websocket.js";
+import socketHandler from "../../../socket/socketHandler.js";
 import { Modal } from "react-native";
 import axiosInstance from "../../../components/axiosInstance.js";
 import fenToJSON from "../../../components/Board/fenToJSON.js";
